+++ src/main/java/com/example/olibrary/service/GenreService.java
@@ -1,5 +1,6 @@
 package com.example.olibrary.service;
 
+import com.example.olibrary.exceptions.NotFoundException;
 import com.example.olibrary.model.Genre;
 import com.example.olibrary.repository.GenreRepository;
 import lombok.extern.slf4j.Slf4j;
@@ -7,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Optional;
+
 @Slf4j
 @Service
 public class GenreService {
@@ -15,11 +18,15 @@ public class GenreService {
 
     public Genre getGenreById(Long id) {
         log.info("Trying to find genre by id={}", id);
-        return genreRepository.findById(id).orElse(null);
+        Optional<Genre> genre = genreRepository.findById(id);
+        if (genre.isEmpty()) {
+            throw new NotFoundException("Genre not found");
+        }
+        return genre.get();
     }
     public ArrayList<Genre> findGenreByName(String name) {
         log.info("Trying to find genre by name like '{}'", name);
-        return genreRepository.findGenreByName(name);
+        return  new ArrayList<>(genreRepository.findGenreByName(name));
     }
 
     public Genre saveGenre(Genre genre) {
@@ -29,6 +36,10 @@ public class GenreService {
 
     public void deleteGenreById(Long id) {
         log.info("Trying to delete by id={}", id);
+        Optional<Genre> genre = genreRepository.findById(id);
+        if (genre.isEmpty()) {
+            throw new NotFoundException("Genre not found");
+        }
         genreRepository.deleteById(id);
     }
 }
