+++ src/main/java/com/example/olibrary/service/UserService.java
@@ -1,10 +1,14 @@
 package com.example.olibrary.service;
 
+import com.example.olibrary.exceptions.NotFoundException;
 import com.example.olibrary.model.User;
 import com.example.olibrary.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
 @Slf4j
 @Service
 public class UserService {
@@ -13,7 +17,11 @@ public class UserService {
 
     public User getUserById(Long id) {
         log.info("Trying to get user by id={}", id);
-        return userRepository.findById(id).orElse(null);
+        Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()) {
+            throw new NotFoundException("Regal not found");
+        }
+        return user.get();
     }
 
     public User getUserByUserName(String username) {
@@ -28,6 +36,10 @@ public class UserService {
 
     public void deleteUserById(Long id) {
         log.info("Trying to delete user by id={}", id);
+        Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()) {
+            throw new NotFoundException("Regal not found");
+        }
         userRepository.deleteById(id);
     }
 }
